BACKEND/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg[binary]==3.2.6
python-decouple==3.8
djangorestframework-simplejwt==5.5.1
django-filter==23.5
drf-spectacular==0.26.5
Pillow==10.4.0
